mig database

+++ cronjob/cron.py
@@ -1,7 +1,6 @@
 import json
 import requests
-import urllib2
-import psycopg2
+import MySQLdb
 import time
 import uuid
 from datetime import datetime
@@ -41,11 +40,10 @@ def get_hotel_list(cursor):
         str_delete_query = "DELETE from hotel"
         cursor.execute(str_delete_query)
 
-        str_insert_query = """INSERT INTO hotel (uuid, hotel_id, hotel_organization_id, hotel_language, hotel_name, hotel_date_modified, hotel_is_active, hotel_deleted, hotel_date_created, hotel_created_by, hotel_modified_by, hotel_network_id, last_update, hotel_energy_defaults_on, hotel_equipment_defaults_on, hotel_humidity_on, hotel_deployment_stage, hotel_screen_settings_on, hotel_default_size_large, hotel_message_degrees_show, hotel_message_screen_contrast, hotel_default_energy_profile, hotel_default_equipment_profile, hotel_presets_on, hotel_pms_integration_on, hotel_additional_network_on, hotel_remote_temperature_control_on, hotel_multi_factor_authentication_on, hotel_btu_average, hotel_kwh_rate, hotel_scheduler_on, hotel_ei_on, hotel_ei_on_date, hotel_vip_ctrl_on, hotel_thermostat_ctrl_on, hotel_ei_saving_screen) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"""
+        str_insert_query = """INSERT INTO hotel (hotel_id, hotel_organization_id, hotel_language, hotel_name, hotel_date_modified, hotel_is_active, hotel_deleted, hotel_date_created, hotel_created_by, hotel_modified_by, hotel_network_id, last_update, hotel_energy_defaults_on, hotel_equipment_defaults_on, hotel_humidity_on, hotel_deployment_stage, hotel_screen_settings_on, hotel_default_size_large, hotel_message_degrees_show, hotel_message_screen_contrast, hotel_default_energy_profile, hotel_default_equipment_profile, hotel_presets_on, hotel_pms_integration_on, hotel_additional_network_on, hotel_remote_temperature_control_on, hotel_multi_factor_authentication_on, hotel_btu_average, hotel_kwh_rate, hotel_scheduler_on, hotel_ei_on, hotel_ei_on_date, hotel_vip_ctrl_on, hotel_thermostat_ctrl_on, hotel_ei_saving_screen) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"""
         arr_insert_query = list()
         for hotel in g_arr_hotel:
             each_hotel = list()
-            each_hotel.append(isNull(str(uuid.uuid4())))
             each_hotel.append(isNull(str(hotel["hotel_id"])))
             each_hotel.append(isNull(str(hotel["hotel_organization_id"])))
             each_hotel.append(isNull(str(hotel["hotel_language"])))
@@ -108,13 +106,12 @@ def get_room_list(cursor):
                 g_arr_room.append(room)
                 arr_room.append(room)
 
-            str_room_insert_query = """INSERT INTO room (uuid, room_id, room_hotel_id, room_name, room_date_created, room_date_modified, room_deleted, room_created_by, room_modified_by, room_energy_profile_id, room_vip_mode, room_equipment_profile_id, room_pms_enabled, room_pms_profile_id, room_number, room_edi_termostat_location, room_edi_termostat_location_hex_b1, room_edi_termostat_location_hex_b2, room_floor_plan_id, room_floor_plan_pos_x, room_floor_plan_pos_y, room_floor_number, room_network_id, room_humidity_allowed, room_energy_report_running, room_cooler_counter, room_heater_counter, room_system_runtime_counter, room_system_runtime_counter_last_value, room_cooler_counter_last_value, room_heater_counter_last_value, room_occupancy_flag, room_hw, room_sw, svn_pic, svn_radio, active_sensor, roomp_termostat_id, roomp_occupancy, roomp_temperature, roomp_thermostat, roomp_heat_setpoint, roomp_cool_setpoint, roomp_auto_setpoint, roomp_system_status_1, roomp_system_status_2, roomp_recovery_time, roomp_max_setpoint_heat, roomp_min_setpoint_cool) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"""
+            str_room_insert_query = """INSERT INTO room (room_id, room_hotel_id, room_name, room_date_created, room_date_modified, room_deleted, room_created_by, room_modified_by, room_energy_profile_id, room_vip_mode, room_equipment_profile_id, room_pms_enabled, room_pms_profile_id, room_number, room_edi_termostat_location, room_edi_termostat_location_hex_b1, room_edi_termostat_location_hex_b2, room_floor_plan_id, room_floor_plan_pos_x, room_floor_plan_pos_y, room_floor_number, room_network_id, room_humidity_allowed, room_energy_report_running, room_cooler_counter, room_heater_counter, room_system_runtime_counter, room_system_runtime_counter_last_value, room_cooler_counter_last_value, room_heater_counter_last_value, room_occupancy_flag, room_hw, room_sw, svn_pic, svn_radio, active_sensor, roomp_termostat_id, roomp_occupancy, roomp_temperature, roomp_thermostat, roomp_heat_setpoint, roomp_cool_setpoint, roomp_auto_setpoint, roomp_system_status_1, roomp_system_status_2, roomp_recovery_time, roomp_max_setpoint_heat, roomp_min_setpoint_cool) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"""
             arr_room_insert_query = list()
             str_alert_insert_query = """INSERT INTO roomalerts (id, alert_type, cms_room_id) VALUES (%s, %s, %s)"""
             arr_alert_insert_query = list()
             for room in arr_room:
                 each_room = list()
-                each_room.append(isNull(str(uuid.uuid4())))
                 each_room.append(isNull(str(room["room_id"])))
                 each_room.append(isNull(str(room["room_hotel_id"])))
                 each_room.append(isNull(str(room["room_name"])))
@@ -183,7 +180,7 @@ def get_report_list(cursor):
         str_report_delete_query = "DELETE from reports"
         cursor.execute(str_report_delete_query)
 
-        str_report_insert_query = """INSERT INTO reports (uuid, room_id, from_time, to_time, comp_runtime, heater_runtime, thermostat_id) VALUES (%s, %s, %s, %s, %s, %s, %s)"""
+        str_report_insert_query = """INSERT INTO reports (room_id, from_time, to_time, comp_runtime, heater_runtime, thermostat_id) VALUES (%s, %s, %s, %s, %s, %s)"""
         arr_report_insert_query = list()
         for room in g_arr_room:
             str_room_id = room['room_id']
@@ -194,7 +191,6 @@ def get_report_list(cursor):
 
             for report in reports:
                 each_report = list()
-                each_report.append(isNull(str(uuid.uuid4())))
                 each_report.append(isNull(str(str_room_id)))
                 each_report.append(isNull(str(str_from)))
                 each_report.append(isNull(str(str_to)))
@@ -204,6 +200,7 @@ def get_report_list(cursor):
                 arr_report_insert_query.append(each_report)
                 g_arr_report.append(each_report)
 
+        print(len(arr_report_insert_query))
         cursor.executemany(str_report_insert_query, arr_report_insert_query)
     except Exception as e:
         print ("exception : get_report_data")
@@ -214,36 +211,35 @@ def get_unoccupied_runtime(cursor):
         str_ocr_delete_query = """DELETE from unoccupied_runtime"""
         cursor.execute(str_ocr_delete_query)
 
-        str_max_allow_query = "SET SESSION max_allowed_packet=500M"
-        cursor.execute(str_max_allow_query)
-
-        str_ocr_insert_query = """INSERT INTO unoccupied_runtime (uuid, s2, c, pt, f, hsp, h, occaux, ttl, occ, dt, t, ob, thermostat_id, hum, csp, sp, s1) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"""
-        arr_ocr_insert_query = list()
         for report in g_arr_report:
-            str_request_url = """https://api.thermostatsolutions.com/v1/thermostats/%s/history?from_date=%s&to_date=%s&access_token=%s""" % (report[6], report[2], report[3], g_strToken)
-            print ("get_unoccupied_runtime : " + str_request_url)
-            unoccupied_runtime = requests.get(str_request_url).json()
+            str_request_url = """https://api.thermostatsolutions.com/v1/thermostats/%s/history?from_date=%s&to_date=%s&access_token=%s""" % (report[5], report[1], report[2], g_strToken)
+            u_a = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/48.0.2564.82 Safari/537.36"
 
-            for data in unoccupied_runtime["data"]:
-                print(data["s2"])
-                print(data["c"])
-                print(data["pt"])
-                print(data["f"])
-                print(data["hsp"])
-                print(data["h"])
-                print(data["occaux"])
-                print(data["ttl"])
-                print(data["occ"])
-                print(data["dt"])
-                print(data["t"])
-                print(data["ob"])
-                print(data["id"])
-                print(data["hum"])
-                print(data["csp"])
-                print(data["sp"])
-                print(data["s1"])
+            recall = 0
+            err_cnt = 0
+            unoccupied_runtime = requests.get(str_request_url, headers={"USER-AGENT": u_a})
+            recall = unoccupied_runtime.status_code
+
+            while (recall != 200):                
+                time.sleep(1)
+                unoccupied_runtime = requests.get(str_request_url, headers={"USER-AGENT": u_a})
+                recall = unoccupied_runtime.status_code
+                err_cnt = err_cnt + 1
+                print (str(recall))
+                if (err_cnt > 20): break                
+            
+            if (err_cnt > 20): continue
+
+            unoccupied_runtime = unoccupied_runtime.json()
+            print(len(unoccupied_runtime["data"]))
+
+            if (len(unoccupied_runtime["data"]) == 0): continue
+
+            str_ocr_insert_query = """INSERT INTO unoccupied_runtime (s2, c, pt, f, hsp, h, occaux, ttl, occ, dt, t, ob, thermostat_id, hum, csp, sp, s1) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"""
+            arr_ocr_insert_query = list()
+
+            for data in unoccupied_runtime["data"]:                
                 arr_unoccupired_runtime = list()
-                arr_unoccupired_runtime.append(isNull(str(uuid.uuid4())))
                 arr_unoccupired_runtime.append(isNull(str(data["s2"])))
                 arr_unoccupired_runtime.append(isNull(str(data["c"])))
                 arr_unoccupired_runtime.append(isNull(str(data["pt"])))
@@ -262,7 +258,10 @@ def get_unoccupied_runtime(cursor):
                 arr_unoccupired_runtime.append(isNull(str(data["sp"])))
                 arr_unoccupired_runtime.append(isNull(str(data["s1"])))
                 arr_ocr_insert_query.append(arr_unoccupired_runtime)
-        cursor.executemany(str_ocr_insert_query, arr_ocr_insert_query)
+
+            print("inserting")
+            cursor.executemany(str_ocr_insert_query, arr_ocr_insert_query)
+            print("inserted")
     except Exception as e:
         print ("exception : get unoccupired data")
         print (str(e))
@@ -271,8 +270,27 @@ def get_occ(cursor):
     try:
         for report in g_arr_report:
             str_request_url = """https://api.thermostatsolutions.com/v1/thermostats/%s/occ-history?from_date=%s&to_date=%s&access_token=%s""" % (report[5], report[1], report[2], g_strToken)
-            print("get_occ : " + str_request_url)
-            occ_list = requests.get(str_request_url).json()
+            
+            u_a = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/48.0.2564.82 Safari/537.36"
+
+            recall = 0
+            err_cnt = 0
+            occ_list = requests.get(str_request_url, headers={"USER-AGENT": u_a})
+            recall = occ_list.status_code
+
+            while (recall != 200):                
+                time.sleep(1)
+                occ_list = requests.get(str_request_url, headers={"USER-AGENT": u_a})
+                recall = occ_list.status_code
+                err_cnt = err_cnt + 1
+                print (str(recall))
+                if (err_cnt > 20): break                
+            
+            if (err_cnt > 20): continue
+
+            occ_list = occ_list.json()
+
+            if (len(occ_list["data"]) == 0): continue
 
             str_occ_delete_query = """DELETE from occhistory WHERE thermostat_id = '%s' AND start_time >= '%s' AND finish_time <= '%s'""" % (report[5], standard_date_format(report[1]), standard_date_format(report[2]))
             cursor.execute(str_occ_delete_query)
@@ -300,10 +318,10 @@ def set_bi_data(cursor):
         print (str(e))
 def main():
     start = time.time()
-    db_conn = psycopg2.connect(
-        host="verdant.c5rdujz93n3m.us-east-1.rds.amazonaws.com",
+    db_conn = MySQLdb.connect(
+        host="verdantdb.mysql.database.azure.com",
         database="verdant",
-        user="postgres",
+        user="admin_verdant@verdantdb",
         password="qweasdzxcasdqwe"
     )
 
@@ -313,7 +331,7 @@ def main():
     get_room_list(cursor)
     get_report_list(cursor)
     get_unoccupied_runtime(cursor)
-    #get_occ(cursor)
+    get_occ(cursor)
     #set_bi_data(cursor)
 
     db_conn.commit()
